schedule block requester

+++ bigtangle-server/src/main/java/net/bigtangle/server/service/schedule/ScheduleMilestoneService.java
@@ -44,9 +44,9 @@ public class ScheduleMilestoneService {
      * read data from table oder by insert time,  use add Block to check again, 
      * if missing previous,  it may request network for the blocks 
      */
-//    @Scheduled(fixedRateString = "${service.updateUnsolideService.rate:30000}")
+    @Scheduled(fixedRateString = "${service.updateUnsolideService.rate:30000}")
     public void updateUnsolideService() {
-        
+        if (scheduleConfiguration.isMilestone_active()) {
             try {
                 logger.debug(" Start ScheduleMilestoneService: ");
                 blockService. deleteOldUnsolidBlock();
@@ -55,7 +55,7 @@ public class ScheduleMilestoneService {
             } catch (Exception e) {
                 logger.warn("updateUnsolideService ", e);
             }
-         
+        }
     }
 
 }
